Display move commands in puzzle when editing moves
